Maze outline coordinate fix
Looks like you're still working on the linearize_trajectory function.
I've corrected the coordinates for the existing code that uses the
output from sections. For future reference, the "bounds" output lists
the xy coordinates of the vertices of the maze sections in
counterclockwise order starting from the bottom left corner.