Bugfix
Fixed some issues with the try/catch statement. Needed to roll back a
trial once there was an error.

Also fixed a issue I missed earlier -- the script was capturing when the
mouse was on the left/right arm right before it was picked up. This
caused dimensional mismatch during concatenation.

Lastly, added .section field to data struct. This is the same as
sect(:,2) from getsection.m