Multiple day image registration
Added function to register one base file across multiple sessions. Note
that previous commits in this branch modified the image_registerX
function such that it now  saves the mat file with the date to which the
base is registered to. This may cause errors when registering new files.
To fix this, delete the RegistrationInfoX.mat file and run
image_register_simple again.